move cards test to another file
