+++ tests/base.py
@@ -5,7 +5,14 @@ from urllib.parse import parse_qsl, urlencode, urlparse, urlunparse
 from flask_testing import TestCase
 
 from gatovid.app import app
-from gatovid.create_db import db_reset, db_test_data
+from gatovid.create_db import (
+    GENERIC_USERS_EMAIL,
+    GENERIC_USERS_NAME,
+    GENERIC_USERS_PASSWORD,
+    NUM_GENERIC_USERS,
+    db_reset,
+    db_test_data,
+)
 from gatovid.exts import db, socket
 
 
@@ -138,6 +145,16 @@ class WsTestClient(GatovidTestClient):
     clients = []
     matchmaking_delay = 0.0
 
+    users_data = [
+        {
+            "email": GENERIC_USERS_EMAIL.format(i),
+            "password": GENERIC_USERS_PASSWORD,
+        }
+        for i in range(NUM_GENERIC_USERS)
+    ]
+
+    player_names = [GENERIC_USERS_NAME.format(i) for i in range(NUM_GENERIC_USERS)]
+
     def create_app(self):
         self.app = super().create_app()
         return self.app
@@ -224,3 +241,26 @@ class WsTestClient(GatovidTestClient):
         """
 
         time.sleep(self.pause_timeout * 1.2)
+
+    def create_game(self, players=6):
+        clients = []
+        for i in range(players):
+            clients.append(self.create_client(self.users_data[i]))
+
+        # Creamos la partida
+        callback_args = clients[0].emit("create_game", callback=True)
+
+        received = clients[0].get_received()
+        _, args = self.get_msg_in_received(received, "create_game", json=True)
+        code = args["code"]
+
+        # Unimos a los clientes a la partida
+        for client in clients[1:]:
+            callback_args = client.emit("join", code, callback=True)
+            self.assertNotIn("error", callback_args)
+
+        # Empezamos la partida
+        callback_args = clients[0].emit("start_game", callback=True)
+        self.assertNotIn("error", callback_args)
+
+        return clients, code
